✨ Feat: 리프레시 토큰 로직 수정 #56

+++ src/libs/auth-axios.ts
@@ -1,7 +1,11 @@
-import type { AxiosError } from 'axios';
+import type { AxiosError, AxiosRequestConfig } from 'axios';
 import axios from 'axios';
 
 import { API_PATHS } from '@/constants/api';
+
+export interface CustomAxiosRequestConfig extends AxiosRequestConfig {
+  _retry?: boolean;
+}
 
 const axiosConfig = {
   baseURL: process.env.NEXT_PUBLIC_API_AUTH_URL,
@@ -27,27 +31,32 @@ axiosAuthInstance.interceptors.request.use(
 axiosAuthInstance.interceptors.response.use(
   (response) => response,
   async (error: AxiosError) => {
-    const originalConfig = error.config;
+    const originalConfig = error.config as CustomAxiosRequestConfig;
+    // eslint-disable-next-line no-underscore-dangle
+    if (error.response && error.response.status === 401 && originalConfig && !originalConfig._retry) {
+      // eslint-disable-next-line no-underscore-dangle
+      originalConfig._retry = true;
 
-    if (error.response && error.response.status === 401 && originalConfig) {
       const refresh = localStorage.getItem('refreshToken');
 
-      try {
-        const { data } = await axiosAuthInstance.post<{ access: string }>(`${API_PATHS.AUTH.TOKEN.REFRESH.POST()}`, { refresh });
+      if (refresh) {
+        try {
+          const { data } = await axiosAuthInstance.post<{ access: string }>(`${API_PATHS.AUTH.TOKEN.REFRESH.POST()}`, { refresh });
 
-        if (data?.access) {
-          localStorage.setItem('accessToken', data.access);
+          if (data?.access) {
+            localStorage.setItem('accessToken', data.access);
 
-          if (originalConfig.headers) {
-            originalConfig.headers.set('Authorization', `Bearer ${data.access}`);
+            if (originalConfig.headers) {
+              originalConfig.headers.Authorization = `Bearer ${data.access}`;
+            }
+
+            return axiosAuthInstance(originalConfig);
           }
-
-          return axiosAuthInstance(originalConfig);
+        } catch (refreshError) {
+          // TODO:로그아웃 처리?
+          localStorage.clear();
+          window.location.href = '/login';
         }
-      } catch (refreshError) {
-        // TODO:로그아웃 처리 ?
-        localStorage.clear();
-        window.location.href = '/login';
       }
     }
